tests/requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
astroid==2.12.10
async-generator==1.10
attrs==22.1.0
by==0.0.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
dill==0.3.5.1
h11==0.13.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
outcome==1.2.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pylint==2.15.3
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-rerunfailures==10.2
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.11.4
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
wsproto==1.2.0
